requirements.txt
PyYAML==3.13
datapackage-pipelines==1.7.1
pyparsing==2.2.0
